Move metrics list to separate file
