+++ internal/model/metrics_list.go
@@ -1,0 +1,8 @@
+package model
+
+var GaugeMetrics = []string{
+	"Alloc", "BuckHashSys", "Frees", "GCCPUFraction", "GCSys", "HeapAlloc", "HeapIdle",
+	"HeapInuse", "HeapObjects", "HeapReleased", "HeapSys", "LastGC", "Lookups", "MCacheInuse",
+	"MCacheSys", "MSpanInuse", "MSpanSys", "Mallocs", "NextGC", "NumForcedGC", "NumGC", "OtherSys",
+	"PauseTotalNs", "StackInuse", "StackSys", "Sys", "TotalAlloc",
+}
