+++ internal/model/model.go
@@ -8,13 +8,6 @@ type Metric struct {
 
 type Data map[string]map[string]Metric
 
-var GaugeMetrics = []string{
-	"Alloc", "BuckHashSys", "Frees", "GCCPUFraction", "GCSys", "HeapAlloc", "HeapIdle",
-	"HeapInuse", "HeapObjects", "HeapReleased", "HeapSys", "LastGC", "Lookups", "MCacheInuse",
-	"MCacheSys", "MSpanInuse", "MSpanSys", "Mallocs", "NextGC", "NumForcedGC", "NumGC", "OtherSys",
-	"PauseTotalNs", "StackInuse", "StackSys", "Sys", "TotalAlloc",
-}
-
 const HTMLTemplateString = `
 <!DOCTYPE html>
 <html>
